@@ -144,8 +144,8 @@
     "opentype.js": "1.1.0",
     "prettier": "1.14.3",
     "prop-types": "15.7.2",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
     "react-native-web": "0.11.4",
     "react-native-web-image-loader": "0.0.6",
     "react-scripts": "3.0.1",
