@@ -51,8 +51,8 @@
     "jsdom": "9.12.0",
     "lodash.merge": "4.6.2",
     "prettier": "1.14.3",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
     "react-native": "0.59.10",
     "react-test-renderer": "16.6.3",
     "webpack": "4.30.0",
