@@ -53,9 +53,9 @@
     "eslint": "5.9.0",
     "jest": "24.8.0",
     "prettier": "1.14.3",
-    "react": "16.8.6",
+    "react": "16.9.0",
     "react-art": "16.6.3",
-    "react-dom": "16.8.6",
+    "react-dom": "16.9.0",
     "react-native": "0.59.10",
     "react-native-web": "0.11.4",
     "react-test-renderer": "16.6.3",
