@@ -58,8 +58,8 @@
     "jest": "24.8.0",
     "jest-styled-components": "6.3.1",
     "prettier": "1.14.3",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
     "react-native": "0.59.10",
     "react-test-renderer": "16.6.3",
     "webpack": "4.30.0",
