@@ -52,7 +52,7 @@ module.exports = variables => ({
     jest: "23.3.0",
     prettier: "1.14.3",
     react: "16.6.3",
-    "react-dom": "16.8.6",
+    "react-dom": "16.9.0",
     "react-native": "0.59.10",
     "react-test-renderer": "16.6.3",
     webpack: "4.30.0",
